requirements.txt
Django==2.2
envparse==0.2.0
psycopg2-binary==2.8.4
dj-database-url==0.5.0
django-tqdm==1.0.0
requests==2.23.0
django-filter==2.2.0
djangorestframework==3.11.0
